requirements.txt
feedparser==6.0.11
requests==2.32.3
beautifulsoup4==4.12.3
lxml==5.3.0
google-genai==1.10.0
python-dateutil==2.9.0
python-dotenv==1.0.1
trafilatura==2.0.0
